feat: add simple confirmation form for delete operations

+++ src/CollectIt.Admin/client/src/components/pages/images/EditImage.tsx
@@ -51,12 +51,14 @@ const EditImage = () => {
     }
 
     const deleteImage = () => {
-        ImagesService.deleteImageByIdAsync(imageId).then(() => {
-            alert('Image deleted successfully');
-            nav('/images');
-        }).catch(x => {
-            alert(x.message);
-        });
+        if (window.confirm('Delete image?')) {
+            ImagesService.deleteImageByIdAsync(imageId).then(() => {
+                alert('Image deleted successfully');
+                nav('/images');
+            }).catch(x => {
+                alert(x.message);
+            });
+        }
     }
 
     return (
